+++ RASPCompiler/src/cup/parser.cup
@@ -7,7 +7,7 @@ import java_cup.runtime.Symbol;
 
 import sk.tuke.emustudio.rasp.compiler.tree.*;
 import sk.tuke.emustudio.rasp.memory.RASPInstruction;
-import sk.tuke.emustudio.rasp.memory.RASPInstructionImpl;
+import sk.tuke.emustudio.rasp.memory.impl.RASPInstructionImpl;
 
 import java.util.List;
 import java.util.Objects;
