+++ RASPCompiler/src/sk/tuke/emustudio/rasp/compiler/tree/Program.java
@@ -10,7 +10,7 @@ import java.util.Collections;
 import java.util.List;
 import sk.tuke.emustudio.rasp.compiler.CompilerOutput;
 import sk.tuke.emustudio.rasp.memory.NumberMemoryItem;
-import sk.tuke.emustudio.rasp.memory.RASPInstructionImpl;
+import sk.tuke.emustudio.rasp.memory.impl.RASPInstructionImpl;
 
 /**
  *
