+++ RASPMemory/src/sk/tuke/emustudio/rasp/memory/impl/RASPInstructionImpl.java
@@ -3,7 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sk.tuke.emustudio.rasp.memory;
+package sk.tuke.emustudio.rasp.memory.impl;
+
+import sk.tuke.emustudio.rasp.memory.RASPInstruction;
 
 public class RASPInstructionImpl implements RASPInstruction {
